+++ index.d.ts
@@ -1,0 +1,4 @@
+declare module LNote {
+    const version: string
+    const root: string
+}