+++ test.mjs
@@ -3,8 +3,10 @@ import fs from 'fs'
 // import { createApp, createSSRApp } from 'vue'
 // import { compileTemplate } from '@vue/compiler-sfc'
 // import { compile } from 'vue-template-compiler'
-import { createSSRApp } from 'vue'
+import { createSSRApp, h } from 'vue'
 import { renderToString, ssrRenderComponent } from '@vue/server-renderer'
+import { parse, compileScript, compileTemplate } from '@vue/compiler-sfc'
+// import Test1 from './test1.vue'
 // const path = require('path')
 // const runPath = process.cwd()
 
@@ -45,19 +47,36 @@ async function test() {
         return
     }
     const tmp = fs.readFileSync('test.vue', 'utf-8')
+    /* const a = await import('./test.vue')
+    console.log(a) */
     const m = tmp.match(/<template>([\s\S]*?)<\/template>/)
-    const app = createSSRApp({
+    /* const app = createSSRApp({
         // template: m[1],
         template: tmp,
         data: () => ({message: 'testing', abc: false, hh: '<div>asd</div>'})
+    }) */
+    const tsr = parse(tmp)
+    // console.log(compileScript(tsr.descriptor, {id: ''}))
+    const tms = compileTemplate({
+        source: tsr.descriptor.template.content,
+        // source: tsr,
+        id: 'xxx'
     })
-    const p = await renderToString(app, { comment: false })
+    // console.log(tms)
+    // return
+    const app = createSSRApp({
+        // template: m[1],
+        template: tsr.descriptor.template.content,
+        data: () => ({message: 'testing', abc: false, hh: '<div>asd</div>'}),
+        // components: []
+    })
+    const p = await renderToString(app, {data: () => ({message: 'testing', abc: false, hh: '<div>asd</div>'})})
     console.log(p)
     /* const s = await ssrRenderComponent(tmp)
     console.log(s) */
 }
 
-// test()
+test()
 
 function testFs() {
     const res = fs.readdirSync('asd')
@@ -65,4 +84,4 @@ function testFs() {
     console.log(res)
 }
 
-testFs()
+// testFs()